updated hui_automation
