+++ src/test/java/hui_automation/boratech_tests/DeleteEducationTest.java
@@ -1,4 +1,4 @@
-package hui_automation;
+package hui_automation.boratech_tests;
 
 import java.time.Duration;
 import java.util.List;
@@ -7,6 +7,8 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
+
+import hui_automation.Testkeys;
 
 public class DeleteEducationTest {
 
@@ -52,6 +54,7 @@ public class DeleteEducationTest {
 			System.out.println("Test passed.");
 			System.out.println("School deleted: " + delSchool);
 			System.out.println("Final total rows: " + eduRows.size());
+			Testkeys.pause(3);
 		} catch (Exception e) {
 			System.out.println("Test failed!");
 			System.out.println("Reason: " + e.getMessage());
